dat_deploy/requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
openai==1.16.1
